Add riddle BL

+++ src/main/kotlin/hu/bme/sch/cmsch/model/RiddleEntity.kt
@@ -58,7 +58,7 @@ data class RiddleEntity(
     @property:GenerateInput(type = INPUT_TYPE_NUMBER, order = 6, label = "Sorrend")
     @property:GenerateOverview(columnName = "Sorrend", order = 3, centered = true)
     @property:ImportFormat(ignore = false, columnId = 4, type = IMPORT_LONG)
-    var order: Long = 0,
+    var order: Int = 0,
 
     @Column(nullable = false)
     @JsonView(value = [ Edit::class, Preview::class, FullDetails::class ])
